Add optional Okta support (#202)
* Deploy ui before ui-auth

* Add application endpoint url support

* Add stuff that is not conditional on building okta or not,  all of this can always be built

* Add userpooldomain, which hopefully can be built unconditionally in reagrds to okta

* Add UserPoolClientDomain output that does not need to be conditional

* Correct typo

* Fix typo causing conditional to not work as intended

* Add the conditional userpoolidentityprovider

* Conditionally add okta as a supported idp

* Adding an implicit dependency since i cant do an explicit conditional dependson

* add new variables to ui-src

* Add okta login... the react code is uncoditional so far

* make the login form look nicer

* remoe unused var

* disable the name fields on the profile page

* run prettier

* Be a little smarter about dependencies

* Move to using the cognito hosted ui.  Remove the login and signup containes

* appease the preflight checks

* remove unneeded logic

* Add cognito as an enabled provider

* prettier

* Set the default user pool id to be of a format that compiles

* get local dev working

* Remove a wrapper

* remove the option to specify a default ssm param for okta

* Remove commented out line
