+++ services/ui-auth/serverless.yml
@@ -28,6 +28,8 @@ custom:
   sesSourceEmailAddress: ${ssm:/configuration/${self:custom.stage}/sesSourceEmailAddress~true, ssm:/configuration/default/sesSourceEmailAddress~true, ""}
   attachments_bucket_arn: ${cf:uploads-${self:custom.stage}.AttachmentsBucketArn}
   api_gateway_rest_api_name: ${cf:app-api-${self:custom.stage}.ApiGatewayRestApiName}
+  okta_metadata_url: ${ssm:/configuration/${self:custom.stage}/okta_metadata_url~true, ""}
+  application_endpoint_url: ${cf:ui-${self:custom.stage}.ApplicationEndpointUrl}
 
 resources:
   Conditions:
@@ -41,6 +43,11 @@ resources:
         - Fn::Equals:
             - ""
             - ${self:custom.iamPermissionsBoundaryPolicy}
+    BackWithOkta:
+      Fn::Not:
+        - Fn::Equals:
+            - ""
+            - ${self:custom.okta_metadata_url}
   Resources:
     CognitoUserPool:
       Type: AWS::Cognito::UserPool
@@ -77,6 +84,52 @@ resources:
         ExplicitAuthFlows:
           - ADMIN_NO_SRP_AUTH
         GenerateSecret: false
+        AllowedOAuthFlows:
+          - implicit
+        AllowedOAuthFlowsUserPoolClient: true
+        AllowedOAuthScopes:
+          - email
+          - openid
+          - aws.cognito.signin.user.admin # Needed to call Auth.currentUserInfo()
+        CallbackURLs:
+          - ${self:custom.application_endpoint_url}
+          - http://localhost:3000/
+        DefaultRedirectURI: ${self:custom.application_endpoint_url}
+        LogoutURLs:
+          - ${self:custom.application_endpoint_url}
+          - http://localhost:3000/
+        SupportedIdentityProviders:
+          - COGNITO
+          - Fn::If:
+              - BackWithOkta
+              - !Ref OktaUserPoolIdentityProvider
+              - !Ref AWS::NoValue
+    UserPoolDomain:
+      Type: AWS::Cognito::UserPoolDomain
+      Properties:
+        Domain:
+          Fn::Join:
+            - ""
+            - - ${self:custom.stage}-login-
+              - Ref: CognitoUserPoolClient
+        UserPoolId:
+          Ref: CognitoUserPool
+    OktaUserPoolIdentityProvider:
+      Condition: BackWithOkta
+      Type: AWS::Cognito::UserPoolIdentityProvider
+      Properties:
+        AttributeMapping:
+          email: http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress
+          given_name: http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname
+          family_name: http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname
+        IdpIdentifiers:
+          - IdpIdentifier
+        ProviderDetails:
+          MetadataURL: ${self:custom.okta_metadata_url}
+        ProviderName: Okta
+        ProviderType: SAML
+        UserPoolId:
+          Ref: CognitoUserPool
     CognitoIdentityPool:
       Type: AWS::Cognito::IdentityPool
       Properties:
@@ -143,6 +196,8 @@ resources:
       Value: !Ref CognitoUserPool
     UserPoolClientId:
       Value: !Ref CognitoUserPoolClient
+    UserPoolClientDomain:
+      Value: !Sub ${UserPoolDomain}.auth.${AWS::Region}.amazoncognito.com
     IdentityPoolId:
       Value: !Ref CognitoIdentityPool
     Region:
