+++ services/ui-src/.env
@@ -1,11 +1,13 @@
 LOCAL_LOGIN=true
 API_REGION=us-east-1
-API_URL=https://c1ng7kke96.execute-api.us-east-1.amazonaws.com/dev
-COGNITO_REGION=us-east-1
-COGNITO_IDENTITY_POOL_ID=us-east-1:6ab3f428-6879-4710-8242-bee2c31654bf
-COGNITO_USER_POOL_ID=us-east-1_rdUFtrS9s
-COGNITO_CLIENT_ID=352mnd3esms3mms1ng256mbvda
+API_URL=placeholder
+COGNITO_REGION=placeholder
+COGNITO_IDENTITY_POOL_ID=placeholder
+COGNITO_USER_POOL_ID=us-east-1_placeholder
+COGNITO_USER_POOL_CLIENT_ID=placeholder
+COGNITO_USER_POOL_CLIENT_DOMAIN=placeholder
+COGNITO_REDIRECT_SIGNIN=placeholder
+COGNITO_REDIRECT_SIGNOUT=placeholder
+S3_ATTACHMENTS_BUCKET_REGION=placeholder
+S3_ATTACHMENTS_BUCKET_NAME=placeholder
 S3_LOCAL_ENDPOINT=http://localhost:4569
-S3_ATTACHMENTS_BUCKET_REGION=us-east-1
-S3_ATTACHMENTS_BUCKET_NAME=local-uploads
-
