+++ services/ui-src/serverless.yml
@@ -31,11 +31,13 @@ custom:
   cognito_region: ${cf:ui-auth-${self:custom.stage}.Region}
   cognito_identity_pool_id: ${cf:ui-auth-${self:custom.stage}.IdentityPoolId}
   cognito_user_pool_id: ${cf:ui-auth-${self:custom.stage}.UserPoolId}
-  cognito_client_id: ${cf:ui-auth-${self:custom.stage}.UserPoolClientId}
+  cognito_user_pool_client_id: ${cf:ui-auth-${self:custom.stage}.UserPoolClientId}
+  cognito_user_pool_client_domain: ${cf:ui-auth-${self:custom.stage}.UserPoolClientDomain}
   s3_attachments_bucket_region: ${cf:uploads-${self:custom.stage}.Region}
   s3_attachments_bucket_name: ${cf:uploads-${self:custom.stage}.AttachmentsBucketName}
   ui_s3_bucket_name: ${cf:ui-${self:custom.stage}.S3BucketName}
   ui_cloudfront_distribution_id: ${cf:ui-${self:custom.stage}.CloudFrontDistributionId}
+  application_endpoint_url: ${cf:ui-${self:custom.stage}.ApplicationEndpointUrl}
   s3Sync:
     - bucketName: ${self:custom.ui_s3_bucket_name}
       localDir: ./build
@@ -59,7 +61,10 @@ custom:
         export COGNITO_REGION=${self:custom.cognito_region}
         export COGNITO_IDENTITY_POOL_ID=${self:custom.cognito_identity_pool_id}
         export COGNITO_USER_POOL_ID=${self:custom.cognito_user_pool_id}
-        export COGNITO_CLIENT_ID=${self:custom.cognito_client_id}
+        export COGNITO_USER_POOL_CLIENT_ID=${self:custom.cognito_user_pool_client_id}
+        export COGNITO_USER_POOL_CLIENT_DOMAIN=${self:custom.cognito_user_pool_client_domain}
+        export COGNITO_REDIRECT_SIGNIN=${self:custom.application_endpoint_url}
+        export COGNITO_REDIRECT_SIGNOUT=${self:custom.application_endpoint_url}
         export S3_ATTACHMENTS_BUCKET_REGION=${self:custom.s3_attachments_bucket_region}
         export S3_ATTACHMENTS_BUCKET_NAME=${self:custom.s3_attachments_bucket_name}
         export LOCAL_LOGIN=false
