+++ services/ui-src/src/config.js
@@ -13,7 +13,10 @@ export default {
   cognito: {
     REGION: window._env_.COGNITO_REGION,
     USER_POOL_ID: window._env_.COGNITO_USER_POOL_ID,
-    APP_CLIENT_ID: window._env_.COGNITO_CLIENT_ID,
+    APP_CLIENT_ID: window._env_.COGNITO_USER_POOL_CLIENT_ID,
+    APP_CLIENT_DOMAIN: window._env_.COGNITO_USER_POOL_CLIENT_DOMAIN,
     IDENTITY_POOL_ID: window._env_.COGNITO_IDENTITY_POOL_ID,
+    REDIRECT_SIGNIN: window._env_.COGNITO_REDIRECT_SIGNIN,
+    REDIRECT_SIGNOUT: window._env_.COGNITO_REDIRECT_SIGNOUT,
   },
 };
