+++ services/ui/serverless.yml
@@ -232,6 +232,22 @@ resources:
                     - /*
         Roles:
           - !Ref DeliveryRole
+    ApplicationEndpointUrl:
+      Type: AWS::SSM::Parameter
+      Properties:
+        Name: /${self:custom.stage}/ui/application_endpoint
+        Type: String
+        Value:
+          Fn::If:
+            - CreateCustomCloudFrontDomain
+            - https://${env:CLOUDFRONT_DOMAIN_NAME, ""}/
+            - Fn::Join:
+                - ""
+                - - https://
+                  - Fn::GetAtt:
+                      - CloudFrontDistribution
+                      - DomainName
+                  - "/"
   Outputs:
     S3BucketName:
       Value: !Ref S3Bucket
@@ -243,3 +259,5 @@ resources:
       Value: !GetAtt Firehose.Arn
     CloudFrontWebAclArn:
       Value: !GetAtt CloudFrontWebAcl.Arn
+    ApplicationEndpointUrl:
+      Value: !GetAtt ApplicationEndpointUrl.Value
